[FIX] payment_mollie_official: correctly evaluate payment acquirer conditions
Case:
1. Configure Mollie and add an API test key.
2. Make sure the payment method Mollie is not set as active (archived) in the payment methods
3. Create a webshop order and continue to https://mollie.odoo-cloud.nl/shop/payment

You'll get a traceback as the t-value "0" was not correctly evaluated causing it trying to render the Mollie payment method its icons. As mollie_aq was empty it was not being able to access the record and values causing in the XML record giving a view traceback to the user.
By setting it to False first we can do a safe check to see if it is not False and then render the XML view with all Mollie payment icons (if active).