[FIX] payment_mollie_official: close directUrl statement
Without this closure the Odoo webpage would give you a 404 page not found on the intermediate page with a traceback such as:
  File "/opt/odoo/custom/mollie-odoo/payment_mollie_official/controllers/__init__.py", line 3, in <module>
    from . import main
File "/opt/odoo/custom/mollie-odoo/payment_mollie_official/controllers/main.py", line 70
    "metadata": {

The confusing part was that a redirect happened but no error was logged, even though this code was invalid. This is now fixed.